refactor(ai): extract prepared compaction request
