+++ packages/ai/src/transports/openai-responses-transport.ts
@@ -37,12 +37,8 @@ import {
   getCompat,
 } from "./openai-transport-params.js";
 
-export {
-  createAzureOpenAIResponsesTransportStreamFn,
-  createOpenAIResponsesTransportStreamFn,
-  buildOpenAIResponsesReasoningReplayMetadata,
-};
-export { requestPreparedOpenAIResponsesCompaction } from "./openai-responses-client.js";
+export { createAzureOpenAIResponsesTransportStreamFn, createOpenAIResponsesTransportStreamFn };
+export { requestPreparedOpenAIResponsesCompaction } from "./openai-responses-compact-request.js";
 export { captureOpenAIResponsesCompaction } from "./openai-responses-compaction-replay.js";
 
 const responsesTesting = {
